20170713 - 오늘 마지막

+++ turnup_fridger/src/kr/co/turnup_fridger/service/impl/RecipeServiceImpl.java
@@ -74,9 +74,7 @@ public class RecipeServiceImpl implements RecipeService{
 		public void updateRecipeIrdnt(Map<String, List> recipeIrdnt) throws NoneRecipeException {
 			//재거할 목록
 			List<Integer> removeIrdntList = recipeIrdnt.get("removeIrdntList");	//삭제할 재료
-			if(removeIrdntList == null || removeIrdntList.size() <= 0 || !removeIrdntList.isEmpty()){
-				//널일때는 아무것도 하지 않음 
-			}else{	
+			if(removeIrdntList != null && removeIrdntList.size()> 0 && !removeIrdntList.isEmpty()){
 				for(int i=0;i<removeIrdntList.size();i++){
 					if(removeIrdntList.get(i) != null )
 					irdntDao.deleteRecipeIrdnt(removeIrdntList.get(i));
@@ -85,10 +83,8 @@ public class RecipeServiceImpl implements RecipeService{
 			
 			//추가할 목록
 			List<RecipeIrdnt> addIrdntList = recipeIrdnt.get("addIrdntList");	//추가할 재료
-			System.out.println("addIrdntList 널첵:"+addIrdntList);
-			if(addIrdntList == null || addIrdntList.size() <= 0 || addIrdntList.isEmpty() ){
-				//널일때는 아무것도 하지 않음 
-			}else{
+//			System.out.println("addIrdntList 널첵:"+addIrdntList);
+			if(addIrdntList != null && addIrdntList.size() > 0 && !addIrdntList.isEmpty()){
 				for(int i=0;i<addIrdntList.size();i++){
 					if(addIrdntList.get(i) != null )
 						irdntDao.insertRecipeIrdnt(addIrdntList.get(i));
@@ -99,19 +95,21 @@ public class RecipeServiceImpl implements RecipeService{
 		@Override
 		public void updateRecipeCrse(Map<String, List> recipeCrse) throws NoneRecipeException {
 
-			//과정
+			//재거할 목록
 			List<Map> removeCrseList = recipeCrse.get("removeCrseList");	//삭제할 과정(recipeId-cookingNo)
-			if(removeCrseList != null || !removeCrseList.isEmpty()){
-			for(int i=0;i<removeCrseList.size();i++){
-				crseDao.deleteRecipeCrse((int)(removeCrseList.get(i).get("recipeId")), (int)(removeCrseList.get(i).get("cookingNo")));
+			if(removeCrseList != null && !removeCrseList.isEmpty() && removeCrseList.size() > 0){
+				for(int i=0;i<removeCrseList.size();i++){
+					crseDao.deleteRecipeCrse((int)(removeCrseList.get(i).get("recipeId")), (int)(removeCrseList.get(i).get("cookingNo")));
+				}
 			}
-			}
-			
+			//추가할 목록
 			List<RecipeCrse> addCrseList = recipeCrse.get("addCrseList");	//추가할 과정
-			if(addCrseList != null || !addCrseList.isEmpty()){
-			for(int i=0; i<addCrseList.size();i++){
-				if(addCrseList.get(i) != null )
-					crseDao.insertRecipeCrse(addCrseList.get(i));
+			if(addCrseList != null && !addCrseList.isEmpty() && addCrseList.size() > 0){
+				for(int i=0; i<addCrseList.size();i++){
+					if(addCrseList.get(i) != null )
+						crseDao.selectRecipeCrseById(addCrseList.get(i).getRecipeId());
+						
+						crseDao.insertRecipeCrse(addCrseList.get(i));
 				}
 			}
 		}
